requirements.txt
flask
flask-cors
python-dotenv
requests
urllib3==1.26.6